make headers for gamestates
