+++ src/code/graph.c
@@ -1,16 +1,24 @@
 #include "libc64/malloc.h"
 #include "libc64/sprintf.h"
 #include "libu64/debug.h"
+
+#include "console_logo_state.h"
 #include "buffers.h"
 #include "gfx.h"
 #include "fault.h"
+#include "file_select_state.h"
 #include "line_numbers.h"
+#include "map_select_state.h"
+#include "prenmi_state.h"
 #include "regs.h"
+#include "setup_state.h"
 #include "sys_ucode.h"
+#include "title_setup_state.h"
 #include "terminal.h"
 #include "ucode_disas.h"
 #include "versions.h"
 #include "z64save.h"
+#include "z64play.h"
 
 #include "macros.h"
 #include "global.h"
