+++ src/code/title_setup.c
@@ -1,4 +1,7 @@
+#include "setup_state.h"
+
 #include "global.h"
+#include "console_logo_state.h"
 
 void Setup_InitImpl(SetupState* this) {
     PRINTF(T("ゼルダ共通データ初期化\n", "Zelda common data initialization\n"));
