+++ src/code/z_play.c
@@ -2,6 +2,7 @@
 #include "buffers.h"
 #include "controller.h"
 #include "fault.h"
+#include "file_select_state.h"
 #include "gfx.h"
 #include "gfxalloc.h"
 #include "kaleido_manager.h"
@@ -20,6 +21,7 @@
 #include "sys_math3d.h"
 #include "sys_matrix.h"
 #include "terminal.h"
+#include "title_setup_state.h"
 #include "versions.h"
 #include "zelda_arena.h"
 #include "z64cutscene_flags.h"
