+++ src/code/z_sram.c
@@ -1,3 +1,4 @@
+#include "file_select_state.h"
 #include "global.h"
 #include "terminal.h"
 #include "versions.h"
