+++ src/overlays/gamestates/ovl_file_choose/z_file_nameset.c
@@ -1,3 +1,4 @@
+#include "file_select_state.h"
 #include "file_select.h"
 
 #include "controller.h"
